Handle state where incoming buffer is not valid utf8 as well as line clear backwards and clear all
Incoming buffer may be truncated and not completely valid. This partially solves the problem by ensuring the data is, in fact, at least a valid utf8 string.

The part of the buffer that is unused is prepended to the next buffer that is processed.

This implementation is incomplete. We may have instructions (CSI, OSC or whatever) that are cut off, are valid utf8, but will fail processing of the escape sequences and likely this will cause issues.