Lock the terminal state and not the internal state
To avoid a scenario where the internal state was in the process of being mutated (but not locked as it was being locked/unlocked) while the GUI was asking for data, move the mutex to cover the outer terminalemulator instead.

I don't know if this scenario was an issue, but this makes more sense to me.