Fix Buffer clipping
We were previously having an issue where the buffer itself was being
clipped properly, but the format tags were off causing the stop of the
scroll back to be wrong. This was caused by not draining the removed
lines from the buffer_line_ranges we had saved.

Additionally, the very very top line of the buffer was being not being
displayed. It was there, but not displayed. This also fixes that.