Reverted all SRM changes
Somehow I either confused myself about what CSI ?12 meant, or it has two
meanings. I *thought* it was a terminal echo mode, where the shell
process would switch between echoing the typed characters back to the
terminal emulator or the shell would not echo those back and the TE was
responsible for inserting them in to the buffer.

Apparently CSI ?12 is setting a blinking cursor.