Send key events if in alternate buffer and we mouse scroll
The key idea here is that some programs do not ask for scroll events from the mouse.

If we are in the primary buffer, we have a scroll back in the buffer and we move the visible portion of the text up, *but do not tell the PTY about our actions*. This is entirely on the TE to handle.

However, if we enter the alternate buffer there is no scroll back. If a user scrolls with the mouse, and there are no *scroll mouse event reporting* going on, we will emit a key event of arrow up/down as appropriate to mimic the scrolling.