Do not handle XtMseUrXvt as well as fix clearing code
By itself this code was fine. However, with freminal now having the ability to respond to queries asking if modes are supported we do not want to give the impression that this is supported at all.

We'll let it fall down the not-supported mode code path.

Addtionally, clear_* methods now clear properly. It appears that clear should just remove the character (not remove the thing from the buffer) and apply the current styling to the cell.