ensure that ENV is properly set for running on it's own pty
We were inheriting some of the env from wez (and other TEs) during the
testing processing so far. This hid a bug where the cursor default
position would be a few characters too far over, at least on zsh, MacOS
using ohmyposh for the command line. Setting `COLORTERM` seems to make
it all happy.

Also adding the version of freminal to the env because, why not?