Wrap Testing fix
The degenerate test case passed. However, we failed on an actual terminal. It turns out there was a tiny tiny tiny edge case triggered in my terminal that would cause it to fail.

Adding the "junk" data to the test case would trigger the fault and the test would fail. The actual fault was that we were offsetting the visible line ranges when we should not have. This bug fixes it. Also added a test case that would trigger the bug.